The "Spring Boot 3.2" card doesn't quite make sense, its title cites a minor version target but the chart includes a "patch" section. Incrementing a patch versions will never result in getting to a new minor version. There isn't a conversion factor like 10 patch bumps -> 1 minor bump. A patch version bump and a minor version bump are qualitatively different.
To make this make sense I am updating the title of the card to cite the patch version target.

Additionally, I am improving the display to determine the spring boot version based on the version of spring-core, rather than the version of spring-boot. This is because libraries used by spring boot applications might benefit from the upgrade recipe without themselves taking a dependency on spring-boot itself. These currently fall through the cracks, showing as "complete" even if they are based on a very old version of spring and do need to be upgraded.